requirements.txt
# Core libraries
langchain
llama-index
haystack

# For a backend server (optional but recommended)
fastapi
uvicorn

# Additional tools
requests
python-dotenv

# Optional AI agent frameworks
# (Install if you plan to use them; some may not be on PyPI or may require extra steps)
autogpt
babyagi